% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abstR.R
\name{abstR}
\alias{abstR}
\title{Function to retrive article abstract data}
\usage{
abstR(pmid = NULL, db = "pubmed", service = "Entrez")
}
\arguments{
\item{pmid}{is a list of PubMed ID (PMID) to look for (e.g., 17284678,9997)}

\item{db}{is the data base to look for abstracts (e.g., pmc or pubmed)}

\item{service}{is the service name (e.g., Entrez)}
}
\description{
Retrive the article data from a PubMed ID.
}
\details{
The function uses internally the seRvice function to set an service,
from which the data are retrived. The Entrez service is used by default. The
result of the scRapper function is a XML structured output.
}
\examples{
abstR(puRepmid(scrpR("apoptosis", retmax = 3)))
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\references{
Eric Sayers, The E-utilities In-Depth: Parameters, Syntax and More, 2015, http://www.ncbi.nlm.nih.gov/books/NBK25499/
}
\keyword{PMID}
\keyword{abstract}
\keyword{scraper}

