% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abstR_text.R
\name{abstR_text}
\alias{abstR_text}
\title{Function to retrive article abstract text}
\usage{
abstR_text(abstr = NULL, aggr = TRUE)
}
\arguments{
\item{abstr}{is a set of data from the abstR function}

\item{aggr}{connectate text.}
}
\description{
Retrive the abstract texts.
}
\details{
TBD.
}
\examples{
abstR_text(abstR(puRepmid(scrpR("apoptosis", retmax = 3))))
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\references{
Nolan, D., Temple Lang, 2014. XML and Web Technologies for Data Sciences with R, 1. Springer-Verlag New York.
}
\keyword{PMID}
\keyword{abstract}
\keyword{scraper}

