% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coRpus.R
\name{coRpus}
\alias{coRpus}
\title{Function to process text data}
\usage{
coRpus(source = NULL)
}
\arguments{
\item{source}{is a set abstrats from the abstR function.}
}
\description{
Create a word heatmap from article abstract data.
}
\details{
TBD.
}
\examples{
coRpus()
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\references{
TBD
}
\keyword{heatmap}

