% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_zip.R
\name{fit_zip}
\alias{fit_zip}
\title{Fit Poisson and ZIP distributions}
\usage{
fit_zip(x)
}
\arguments{
\item{x}{a \code{numeric} count vector}
}
\description{
Fit Poisson and Zero-Inflated Poisson distributions to data.
}
\details{
TBD.
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\references{
Lambert D, \emph{Zero-Inflated Poisson regression, with an 
application to defects in manufacturing}. Technometrics. 34(1):1-14.
}
\seealso{
Parameters of Zero-Inflated Poisson are estimated by 
\code{\link[pscl]{zeroinfl}} function.
}
\keyword{models}

