% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mdcr-package.R
\docType{package}
\name{mdcr-package}
\alias{mdc}
\alias{mdcr-package}
\title{Analysis of experiments}
\description{
The package imports and processes various data.
}
\author{
Stefan Roediger, Michal Burdukiewicz.

Maintainer: Stefan Roediger <Stefan.Roediger@b-tu.de>
}
\keyword{package}

