% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/puRepmid.R
\name{puRepmid}
\alias{puRepmid}
\title{Function to retrive data from a scrpR search}
\usage{
puRepmid(scrpRdat = NULL)
}
\arguments{
\item{scrpRdat}{is a scrpR function search result}
}
\description{
Retrive the pmid (PubMed identifier) from a scrpR search.
}
\details{
The Entrez service is used by default. The
result of the scRapper function is a XML structured output.
}
\examples{
puRepmid(scrpR("apoptosis", retmax = 3))
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{scraper}

