% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_aklides.R
\name{read_aklides}
\alias{read_aklides}
\title{Import data from the Aklides platform}
\usage{
read_aklides(file_name)
}
\arguments{
\item{file_name}{is the name of the file}
}
\description{
Imports a single raw data EXCEL file as exported from the Aklides platform.
}
\details{
The raw format means that every column corresponds to a specific
export value of the Aklides platform
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{Aklides}

