% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scrpR.R
\name{scrpR}
\alias{scrpR}
\title{Function to retrive data from a service}
\usage{
scrpR(expr = NULL, retmax = 20, service = "Entrez")
}
\arguments{
\item{expr}{is expression to look for (e.g., H2AX)}

\item{retmax}{is the total number of UIDs from the retrieved set to be shown in the XML output}

\item{service}{is the service name (e.g., Entrez)}
}
\description{
Retrive the pmid (PubMed identifier).
}
\details{
The function uses internally the seRvice function to set an service,
from which the data are retrived. The Entrez service is used by default. The
result of the scRapper function is a XML structured output.
}
\examples{
scrpR("apoptosis", retmax = 3)
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\references{
Eric Sayers, The E-utilities In-Depth: Parameters, Syntax and More, 2015, http://www.ncbi.nlm.nih.gov/books/NBK25499/
}
\keyword{scraper}

