% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seRvice.R
\name{seRvice}
\alias{seRvice}
\title{Helper function to select a service for data retrival}
\usage{
seRvice(sl = "Entrez")
}
\arguments{
\item{sl}{is the service name (e.g., Entrez)}
}
\description{
Sets the options for an R environment to retrive data from a defined source.
}
\details{
This function sets the URL of a scientific web service to be used 
for the data retrival. Several data sources have been build durig the past 
few decades. The Entrez service is one of the most popular.
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\references{
Eric Sayers, The E-utilities In-Depth: Parameters, Syntax and More, 2015, http://www.ncbi.nlm.nih.gov/books/NBK25499/
}
\keyword{service}

