% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelleR.R
\name{spelleR}
\alias{spelleR}
\title{Function to retrive spelling suggestions}
\usage{
spelleR(expr = NULL, db = "pmc", service = "Entrez")
}
\arguments{
\item{expr}{is expression to look for (e.g., H2AX)}

\item{db}{is the data base to look for spelling suggestions (e.g., pmc)}

\item{service}{is the service name (e.g., Entrez)}
}
\description{
Retrive the pmid (PubMed identifier).
}
\details{
The function uses internally the seRvice function to set an service,
from which the data are retrived. The Entrez service is used by default. The 
result of the scRapper function is a XML structured output.
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\references{
Eric Sayers, The E-utilities In-Depth: Parameters, Syntax and More, 2015, http://www.ncbi.nlm.nih.gov/books/NBK25499/
}
\keyword{scraper}

